requirements.txt
fastapi
uvicorn[standard]
proxmoxer
requests
apscheduler
sqlalchemy
python-dotenv
paramiko
